# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Nova Mart Commerce API
  version: 1.0.0
  description: |
    Public HTTP contract for the **Nova Mart** online shopping platform.

    This document is the **source of truth**. It was authored *before* any service
    implementation, and every backend endpoint and every frontend API client method
    is derived from it. Changing behaviour means changing this file first.

    ## Architecture

    Clients never talk to a service directly. All traffic enters through the
    **API Gateway** (`/api/v1/**`), which authenticates the caller, strips any
    client-supplied identity headers, injects trusted ones, and routes to the
    owning service.

    | Path prefix                                   | Owning service       | Owns database     |
    | --------------------------------------------- | -------------------- | ----------------- |
    | `/auth`, `/users`                              | auth-service         | `auth_db`         |
    | `/products`, `/categories`, `/brands`          | product-service      | `product_db`      |
    | `/cart`                                        | cart-service         | `cart_db`         |
    | `/orders`                                      | order-service        | `order_db`        |
    | `/payments`                                    | payment-service      | `payment_db`      |
    | `/inventory`                                   | inventory-service    | `inventory_db`    |
...

...

    ## Response envelope

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      operationId: listProducts
      summary: Browse the catalogue
      description: |
        Public, paginated catalogue search. All filters combine with AND.
        `sort` accepts `field,direction` where direction is `asc` or `desc`.
      security: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: search
          in: query
          description: Free-text match against product name and description
          schema:
            type: string
          example: headphones
        - name: category
          in: query
          description: Category slug
          schema:
            type: string
          example: audio
        - name: brand
          in: query
          description: Brand slug
          schema:
            type: string
          example: aurelia-audio
        - name: minPrice
          in: query
          schema:
            type: number
            format: double
            minimum: 0
        - name: maxPrice
          in: query
          schema:
            type: number
            format: double
            minimum: 0
        - name: featured
          in: query
          description: Restrict to products flagged for the storefront home page
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            default: createdAt,desc
            enum:
              - createdAt,desc
              - price,asc
              - price,desc
              - name,asc
              - name,desc
              - ratingAverage,desc
          example: price,asc
      responses:
        '200':
          description: Paged catalogue
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedProducts'
        '400':
          $ref: '#/components/responses/ValidationError'
    post:
      tags: [Products]
      operationId: createProduct
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: SKU or slug already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Product UUID or URL slug
        schema:
          type: string
        example: aurelia-halo-anc-headphones
    get:
      tags: [Products]
      operationId: getProduct
      summary: Get a single product
      description: Accepts either the product UUID or its slug, so storefront URLs stay readable.
      security: []
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Products]
      operationId: updateProduct
      summary: Update a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Products]
      operationId: deleteProduct
      summary: Delete a product (admin)
      description: Soft-deletes the product so historical order lines remain resolvable.
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/{id}/image:
    parameters:
      - name: id
        in: path
        required: true
        description: Product UUID
        schema:
          type: string
          format: uuid
        example: 8b7d5e62-9f3a-4e8c-b631-7e04f2d3a901
    post:
      tags: [Products]
      operationId: uploadProductImage
      summary: Upload a product image from local device (admin)
      description: |
        Uploads a PNG, JPG, JPEG, or WEBP image file up to 5 MB for the product.
        Persists the image to local storage, saves the relative path in the product database,
        and deletes any superseded uploaded image.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: The image file (JPG, JPEG, PNG, WEBP, max 5 MB)
      responses:
        '200':
          description: Product image uploaded successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductImageUploadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /products/batch:
    post:
      tags: [Products]
      operationId: getProductsByIds
      summary: Resolve many products at once
      description: |
        Bulk lookup used by other services (order-service takes a price and name
        snapshot at checkout time) and by the cart page. Avoids N+1 round trips.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productIds]
              properties:
                productIds:
                  type: array
                  minItems: 1
                  maxItems: 100
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Matching products, in no guaranteed order. Unknown ids are omitted.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # =====================================================================
  #  CATEGORIES AND BRANDS
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT issued by auth-service. Send as `Authorization: Bearer <accessToken>`.'
  parameters:
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      description: Page size
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 12
  responses:
    ValidationError:
      description: The request body or query string failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Request validation failed
            errorCode: VALIDATION_FAILED
            timestamp: '2026-02-14T09:40:12.882Z'
            path: /api/v1/auth/register
            fieldErrors:
              - field: password
                message: Password must be at least 8 characters
              - field: email
                message: Must be a well-formed email address
    Unauthorized:
      description: Missing, malformed or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Authentication is required to access this resource
            errorCode: UNAUTHORIZED
            timestamp: '2026-02-14T09:41:03.114Z'
            path: /api/v1/cart
    Forbidden:
      description: Authenticated, but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: You do not have permission to perform this action
            errorCode: FORBIDDEN
            timestamp: '2026-02-14T09:42:55.900Z'
            path: /api/v1/products
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: Product not found
            errorCode: PRODUCT_NOT_FOUND
            timestamp: '2026-02-14T09:43:30.720Z'
            path: /api/v1/products/8f14e45f-ea8f-4b2a-9d1c-3c2b7a51de90
  schemas:
...
    ApiResponse:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Product retrieved successfully
        data:
          description: Operation-specific payload. Absent for `204` responses.
    ErrorResponse:
      type: object
      required: [success, message, errorCode, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        message:
          type: string
          description: Human-readable, safe to show a shopper. Never contains a stack trace.
        errorCode:
          type: string
          description: Stable machine-readable identifier. Clients branch on this, not on the message text.
          example: PRODUCT_NOT_FOUND
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        fieldErrors:
          type: array
          description: Present only when `errorCode` is `VALIDATION_FAILED`.
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 12
        totalElements:
          type: integer
          format: int64
          example: 24
        totalPages:
          type: integer
          example: 2
        first:
          type: boolean
        last:
          type: boolean

    # ---------- auth ----------
...
    Product:
      type: object
      required: [id, sku, slug, name, price, currency, categoryId, categoryName, imageUrl, active]
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          example: AUR-HALO-BLK
        slug:
          type: string
          example: aurelia-halo-anc-headphones
        name:
          type: string
        shortDescription:
          type: string
        price:
          type: number
          format: double
          description: Current selling price
          example: 18999.00
        compareAtPrice:
          type: number
          format: double
          nullable: true
          description: Original price when the item is discounted. Null when not on offer.
          example: 24999.00
        discountPercent:
          type: integer
          nullable: true
          description: Derived from price and compareAtPrice. Null when not on offer.
          example: 24
        currency:
          type: string
          example: INR
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
        categorySlug:
          type: string
        brandId:
          type: string
          format: uuid
          nullable: true
        brandName:
          type: string
          nullable: true
        imageUrl:
          type: string
          format: uri
        ratingAverage:
          type: number
          format: double
          minimum: 0
          maximum: 5
          example: 4.6
        ratingCount:
          type: integer
          example: 214
        featured:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            description:
              type: string
            images:
              type: array
              items:
                type: string
                format: uri
            specifications:
              type: array
              items:
                type: object
                required: [label, value]
                properties:
                  label:
                    type: string
                  value:
                    type: string
            availability:
              $ref: '#/components/schemas/InventoryItem'
    ProductRequest:
      type: object
      required: [sku, name, description, price, categoryId]
      properties:
        sku:
          type: string
          minLength: 2
          maxLength: 40
        name:
          type: string
          minLength: 2
          maxLength: 180
        shortDescription:
          type: string
          maxLength: 300
        description:
          type: string
          minLength: 10
          maxLength: 5000
        price:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        compareAtPrice:
          type: number
          format: double
          minimum: 0
          nullable: true
        categoryId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
          nullable: true
        imageUrl:
          type: string
          description: Persistent image path or URL
          example: /uploads/products/8b7d5e62-9f3a-4e8c-b631-7e04f2d3a901.webp
        images:
          type: array
          items:
            type: string
        specifications:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              value:
                type: string
        featured:
          type: boolean
          default: false
        active:
          type: boolean
          default: true
        initialStock:
          type: integer
          minimum: 0
          description: |
            Applied on creation only. product-service calls inventory-service to
            seed the stock record, because inventory is a separate bounded context
            and product-service must not write to `inventory_db` directly.
    ProductImageUploadResponse:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
          description: Persistent relative path of the uploaded image file
          example: /uploads/products/8b7d5e62-9f3a-4e8c-b631-7e04f2d3a901.webp
...
    PagedProducts:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          $ref: '#/components/schemas/PageMeta'

    # ---------- cart ----------
...
    InventoryItem:
      type: object
      required: [productId, totalQuantity, reservedQuantity, availableQuantity, inStock]
      properties:
        productId:
          type: string
          format: uuid
        totalQuantity:
          type: integer
          description: Physical units held.
        reservedQuantity:
          type: integer
          description: Units held for orders that are in flight but not yet settled.
        availableQuantity:
          type: integer
          description: totalQuantity minus reservedQuantity. This is what a shopper may buy.
        reorderThreshold:
          type: integer
          description: Availability at or below this level marks the product low stock in the admin dashboard.
          example: 5
        inStock:
          type: boolean
        updatedAt:
          type: string
          format: date-time
...